Massive changelog - Sorted out lots of bits of the API and seperated them from Line.x. Cleaned up the kernel somewhat. Moved CPUID to a set of inline assembly functions using GCC's AT&T Syntax as opposed to the previous use of NASM external functions with the Intel syntax

+++ Kernel/headers/timers/clock.h
@@ -7,6 +7,6 @@
 //Export initialize_system_clock sets up the correct timer and callbacks
 void initialize_system_clock();
 
-unsigned long long get_clock_ticks(); //Number of time the clock has "ticked"
+unsigned long get_clock_ticks(); //Number of time the clock has "ticked"
 
 #endif //_CLOCK_DEF_H_
