+++ Kernel/sources/timers/clock.c
@@ -2,7 +2,7 @@
 #include <panic/panic.h>
 #include "pit.h"
 
-unsigned long long clock_ticks;
+unsigned long clock_ticks;
 
 //When the PIT is triggered call a scheduler tick and add one to the clock ticks (Setup so different timers can be used for the scheduler/clock Etcetera)
 void pit_callback()
@@ -19,7 +19,7 @@ void initialize_system_clock()
 	init_pit(1000);
 }
 
-unsigned long long get_clock_ticks()
+unsigned long get_clock_ticks()
 {
 	return clock_ticks;
 }
